SellerDTO 삭제 및 SellerService, SellerServiceImpl 수정

+++ core/src/main/java/com/triplan/domain/ReservationVO.java
@@ -13,14 +13,14 @@ import java.util.Date;
 @ToString
 @Builder
 public class ReservationVO {
-    private Long resId;
-    private Long memberId;
-    private Long itemId;
-    private Long totalPrice;
-    private Long discountPrice;
-    private Long finalPrice;
+    private Integer resId;
+    private Integer memberId;
+    private Integer itemId;
+    private Integer totalPrice;
+    private Integer discountPrice;
+    private Integer finalPrice;
     private Date resDate;
-    private Long numberOfPerson;
+    private Integer numberOfPerson;
     private String coupon;
     private String name;
     private String phone;
